mysite/requirements.txt
beautifulsoup4==4.9.1
click==7.1.1
click-plugins==1.1.1
Django==2.2
dropbox==11.4.1
firebase==3.0.1
firebase-admin==3.2.1
future==0.18.2
gcloud==0.18.3
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-cloud-core==1.2.0
google-cloud-firestore==1.6.1
google-cloud-storage==1.25.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
html2text==2020.1.16
html5lib==1.0.1
HTMLParser==0.0.2
httplib2==0.16.0
idna==2.6
isort==4.3.21
jws==0.1.3
lazy-object-proxy==1.4.3
lxml==4.6.1
matplotlib==3.2.1
mccabe==0.6.1
mechanize==0.4.5
msedge-selenium-tools==3.141.3
msgpack==0.6.2
multidict==4.7.6
Naked==0.1.31
nmap==0.0.1
numpy==1.18.2
oauth2client==4.1.2
packaging==16.8
pefile==2019.4.18
pexpect==4.2.1
Pillow==7.2.0
pluggy==0.5.1
ply==3.11
protobuf==3.11.2
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.4
pydub==0.24.1
PyGetWindow==0.0.8
PyInstaller==3.6
pylint==2.6.0
pylint-plugin-utils==0.6
PyMsgBox==1.0.8
pynput==1.6.8
pyparsing==2.4.7
Pyrebase4==4.4.1
PyRect==0.1.4
PyScreeze==0.1.26
PySlowLoris==0.1.30
python-dateutil==2.6.0
python-firebase==1.2
python-jwt==2.0.1
python-nmap==0.6.1
PyTweening==1.0.3
pytz==2019.3
pywin32-ctypes==0.2.0
PyYAML==5.3
requests==2.25.1
requests-toolbelt==0.9.1
robobrowser==0.5.3
rsa==4.0
scapy==2.4.3
selenium==3.141.0
shellescape==3.4.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
sseclient==0.0.24
stone==3.2.1
toml==0.10.2
tqdm==4.43.0
uritemplate==3.0.1
urllib3==1.22
virtualenv==20.0.31
webencodings==0.5.1
XlsxWriter==1.2.8

